requirements.txt
absl-py==2.3.0
astunparse==1.6.3
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.59.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.0
h5py==3.14.0
idna==3.10
joblib==1.5.1
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.7.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.1.0
numpy==1.26.4
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==5.29.5
Pygments==2.19.2
pyparsing==3.0.9
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rich==14.0.0
scikit-learn==1.2.2
scipy==1.16.1
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
